Created new StaticDataSource and renmaed existing StaticDataSource to StaticListDataSource
