+++ src/java/com/sarality/app/view/datasource/StaticDataSource.java
@@ -1,26 +1,16 @@
 package com.sarality.app.view.datasource;
 
-import java.util.ArrayList;
-import java.util.List;
-
 /**
- * A ListDataSource with static list of elements
- * 
- * @author abhideep@dothat.in (Abhideep Singh)
+ * A DataSource with static list of elements
  *
  * @param <T> Type of data provided by the data set.
+ * @author sunayna@dothat.in (Sunayna Uberoy)
  */
-public class StaticDataSource<T> implements DataSource<List<T>> {
-  private final List<T> dataset = new ArrayList<T>();
+public class StaticDataSource<T> implements DataSource<T> {
+  private final T data;
 
-  public StaticDataSource(T... elements) {
-    for (T element : elements) {
-      dataset.add(element);
-    }
-  }
-
-  public StaticDataSource(List<T> elementList) {
-    dataset.addAll(elementList);
+  public StaticDataSource(T data) {
+    this.data = data;
   }
 
   @Override
@@ -34,7 +24,8 @@ public class StaticDataSource<T> implements DataSource<List<T>> {
   }
 
   @Override
-  public List<T> getData() {
-    return dataset;
+  public T getData() {
+    return data;
   }
 }
+
